Update q1.cpp
Question-1: A newly-discovered specie is under research. At time t = 0 hours, a scientist puts one hundred specified specie (A0=100) into a favorable growth medium. He wants to compute how many more specie will be after six hours. Assuming exponential growth, the growth constant k for the bacteria is 0.25/hr. The requirement is to create a generic program that runs for an ample range of values of these parameters. Ask the user for the parameters, then you create a more general program, which would work for larger range of values. Write a C++ program for it. [Hint: Include library cmath to use pow( ) function] The amount of specified specie A after an elapsed time t (starting at to=0) is as follows:  A=A0*exp(k*t)