Add files via upload
Acidity or alkalinity of a substance is defined according to the formula pH = –log[H+] where [H+] refers to the hydrogen ion concentration in moles per liter and log is the natural logarithm. Solutions with a pH value of less than 7 are acidic; solutions with a pH value of greater than 7 are basic; solutions with a pH of 7 (such as pure water) are neutral. Suppose that you test your program with apple juice and find that the hydrogen ion concentration is [H+] = 0.0003. Find the pH value for the apple juice and display it. Write a C++ program for it. [Hint: Include library cmath to use log function]