Add files via upload
In a college, some students are in a difficulty. They are running out of time and they have to perform an experiment to prepare a solution that contains certain chemical. They are provided with 250 milliliters of a solution having 14% of that chemical. In this solution, they have to further add another solution that has 54% of that chemical. Take the amount of this solution as 2 liters which should be entered by the user and declare it as an integer. Convert it into milliliters. A programmer is a helper too! You as a programmer should help the students to determine thetotal quantity of chemical in the newly formed combined solution. Write a C++ program for it anddisplay the output. [Hint: An equation will be formed and, you will have to solve it]